+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/application/command/transaction/command/StartSellTransactionCommand.java
@@ -21,12 +21,14 @@ import com.icoin.trading.tradeengine.domain.model.order.OrderBookId;
 import com.icoin.trading.tradeengine.domain.model.portfolio.PortfolioId;
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
 public class StartSellTransactionCommand extends AbstractStartTransactionCommand {
 
-    public StartSellTransactionCommand(TransactionId transactionId, OrderBookId orderbookIdentifier, PortfolioId portfolioIdentifier, long tradeCount, long itemPrice) {
+    public StartSellTransactionCommand(TransactionId transactionId, OrderBookId orderbookIdentifier, PortfolioId portfolioIdentifier, BigDecimal tradeCount, BigDecimal itemPrice) {
         super(transactionId, orderbookIdentifier, portfolioIdentifier, tradeCount, itemPrice);
     }
 }
