+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/domain/events/portfolio/coin/ItemReservationConfirmedForPortfolioEvent.java
@@ -20,6 +20,8 @@ import com.icoin.trading.tradeengine.domain.model.order.OrderBookId;
 import com.icoin.trading.tradeengine.domain.model.portfolio.PortfolioId;
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
@@ -27,19 +29,19 @@ public class ItemReservationConfirmedForPortfolioEvent {
     private PortfolioId portfolioIdentifier;
     private OrderBookId orderBookIdentifier;
     private TransactionId transactionIdentifier;
-    private long amountOfConfirmedItems;
+    private BigDecimal amountOfConfirmedItems;
 
     public ItemReservationConfirmedForPortfolioEvent(PortfolioId portfolioIdentifier,
                                                      OrderBookId orderBookIdentifier,
                                                      TransactionId transactionIdentifier,
-                                                     long amountOfConfirmedItems) {
+                                                     BigDecimal amountOfConfirmedItems) {
         this.portfolioIdentifier = portfolioIdentifier;
         this.orderBookIdentifier = orderBookIdentifier;
         this.transactionIdentifier = transactionIdentifier;
         this.amountOfConfirmedItems = amountOfConfirmedItems;
     }
 
-    public long getAmountOfConfirmedItems() {
+    public BigDecimal getAmountOfConfirmedItems() {
         return amountOfConfirmedItems;
     }
