+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/domain/events/transaction/AbstractTransactionExecutedEvent.java
@@ -19,15 +19,17 @@ package com.icoin.trading.tradeengine.domain.events.transaction;
 
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
 public abstract class AbstractTransactionExecutedEvent {
     private TransactionId transactionIdentifier;
-    private long amountOfItems;
-    private long itemPrice;
+    private BigDecimal amountOfItems;
+    private BigDecimal itemPrice;
 
-    public AbstractTransactionExecutedEvent(TransactionId transactionIdentifier, long amountOfItems, long itemPrice) {
+    public AbstractTransactionExecutedEvent(TransactionId transactionIdentifier, BigDecimal amountOfItems, BigDecimal itemPrice) {
         this.transactionIdentifier = transactionIdentifier;
         this.amountOfItems = amountOfItems;
         this.itemPrice = itemPrice;
@@ -37,11 +39,11 @@ public abstract class AbstractTransactionExecutedEvent {
         return transactionIdentifier;
     }
 
-    public long getAmountOfItems() {
+    public BigDecimal getAmountOfItems() {
         return amountOfItems;
     }
 
-    public long getItemPrice() {
+    public BigDecimal getItemPrice() {
         return itemPrice;
     }
 }
