+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/domain/events/transaction/BuyTransactionCancelledEvent.java
@@ -19,12 +19,14 @@ package com.icoin.trading.tradeengine.domain.events.transaction;
 
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
 public class BuyTransactionCancelledEvent extends AbstractTransactionCancelledEvent {
 
-    public BuyTransactionCancelledEvent(TransactionId transactionIdentifier, long totalAmountOfItems, long amountOfExecutedItems) {
+    public BuyTransactionCancelledEvent(TransactionId transactionIdentifier, BigDecimal totalAmountOfItems, BigDecimal amountOfExecutedItems) {
         super(transactionIdentifier, totalAmountOfItems, amountOfExecutedItems);
     }
 }
