+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/domain/events/transaction/BuyTransactionPartiallyExecutedEvent.java
@@ -19,12 +19,14 @@ package com.icoin.trading.tradeengine.domain.events.transaction;
 
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
 public class BuyTransactionPartiallyExecutedEvent extends AbstractTransactionPartiallyExecutedEvent {
 
-    public BuyTransactionPartiallyExecutedEvent(TransactionId transactionIdentifier, long amountOfExecutedItems, long totalOfExecutedItems, long itemPrice) {
+    public BuyTransactionPartiallyExecutedEvent(TransactionId transactionIdentifier, BigDecimal amountOfExecutedItems, BigDecimal totalOfExecutedItems, BigDecimal itemPrice) {
         super(transactionIdentifier, amountOfExecutedItems, totalOfExecutedItems, itemPrice);
     }
 }
