+++ trade-engine/src/main/java/com/icoin/trading/tradeengine/domain/events/transaction/SellTransactionStartedEvent.java
@@ -20,6 +20,8 @@ import com.icoin.trading.tradeengine.domain.model.order.OrderBookId;
 import com.icoin.trading.tradeengine.domain.model.portfolio.PortfolioId;
 import com.icoin.trading.tradeengine.domain.model.transaction.TransactionId;
 
+import java.math.BigDecimal;
+
 /**
  * @author Jettro Coenradie
  */
@@ -28,8 +30,8 @@ public class SellTransactionStartedEvent extends AbstractTransactionStartedEvent
     public SellTransactionStartedEvent(TransactionId transactionIdentifier,
                                        OrderBookId orderbookIdentifier,
                                        PortfolioId portfolioIdentifier,
-                                       long totalItems,
-                                       long pricePerItem) {
+                                       BigDecimal totalItems,
+                                       BigDecimal pricePerItem) {
         super(transactionIdentifier, orderbookIdentifier, portfolioIdentifier, totalItems, pricePerItem);
     }
 }
